KEBA EVCS: cleanup
- Split impl and interface
- Remove unnecessary 'Alias' channel
